+++ src/interactive_mid_protocols/SigmaProtocolDamgardJurikProduct.cpp
@@ -217,7 +217,7 @@ bool SigmaDJProductSimulator::checkSoundnessParam(biginteger modulus) {
 */
 bool SigmaDJProductSimulator::checkChallengeLength(vector<byte> challenge) {
 	//If the challenge's length is equal to t, return true. else, return false.
-	return (challenge.size() == (t / 8) ? true : false);
+	return ((int) challenge.size() == (t / 8) ? true : false);
 }
 
 /**
@@ -346,7 +346,7 @@ bool SigmaDJProductProverComputation::checkSoundnessParam(biginteger modulus) {
 */
 bool SigmaDJProductProverComputation::checkChallengeLength(vector<byte> challenge) {
 	//If the challenge's length is equal to t, return true. else, return false.
-	return (challenge.size() == (t / 8) ? true : false);
+	return ((int) challenge.size() == (t / 8) ? true : false);
 }
 
 /**
